torch 1.1.0 compatibility for Yolo loss function
Summary: error found during @kevinaer's version testing

Reviewers: kevinaer, mark.kurtz, tuan, dhuang, mgoin

Reviewed By: kevinaer, mark.kurtz

Subscribers: kevinaer, #core

Differential Revision: https://phabricator.neuralmagic.com:8021/D2615

+++ tests/pytorch/utils/test_loss.py
@@ -244,9 +244,9 @@ class TestSSDLossWrapperImpl(LossWrapperTest):
             (
                 torch.cat(
                     (
-                        torch.arange(4)[:, None],
+                        torch.arange(4)[:, None].float(),
                         torch.rand(4, 4),
-                        torch.randint(80, (4, 1)),
+                        torch.randint(80, (4, 1)).float(),
                     ),
                     1,
                 ),
